Moved instance methods for users to separate file. Moved sendByTemplate method to instance method
