+++ server/models/user.js
@@ -305,76 +305,6 @@ UserSchema.pre('save', function(next) {
 
 });
 
-/**
- * Methods
- */
-
-/**
- * HasRole - check if the user has required role
- *
- * @param {String} plainText
- * @return {Boolean}
- * @api public
- */
-UserSchema.methods.hasRole = function(role) {
-  var roles = this.roles;
-  return roles.indexOf('admin') !== -1 || roles.indexOf(role) !== -1;
-};
-
-/**
- * IsAdmin - check if the user is an administrator
- *
- * @return {Boolean}
- * @api public
- */
-UserSchema.methods.isAdmin = function() {
-  return this.roles.indexOf('admin') !== -1;
-};
-
-/**
- * Authenticate - check if the passwords are the same
- *
- * @param {String} plainText
- * @return {Boolean}
- * @api public
- */
-UserSchema.methods.authenticate = function(plainText) {
-  return this.hashPassword(plainText) === this.hashed_password;
-};
-
-/**
- * Make salt
- *
- * @return {String}
- * @api public
- */
-UserSchema.methods.makeSalt = function() {
-  return crypto.randomBytes(16).toString('base64');
-};
-
-/**
- * Hash password
- *
- * @param {String} password
- * @return {String}
- * @api public
- */
-UserSchema.methods.hashPassword = function(password) {
-  if (!password || !this.salt) return '';
-  var salt = new Buffer(this.salt, 'base64');
-  return crypto.pbkdf2Sync(password, salt, 10000, 64, 'sha512').toString('base64');
-};
-
-/**
- * Hide security sensitive fields
- *
- * @returns {*|Array|Binary|Object}
- */
-UserSchema.methods.toJSON = function() {
-  var obj = this.toObject();
-  delete obj.hashed_password;
-  delete obj.salt;
-  return obj;
-};
+UserSchema.methods = require('./instance-methods');
 
 mongoose.model('User', UserSchema);
